requirements.txt
Django==4.1.3
python-dotenv
requests
smtplib
ssl
email